Style: Moved Misc to its own module
